// src/enums/news-categories.ts
export enum NewsCategories {
   BusinessNews = 'BusinessNews',
   GamingNews = 'GamingNews',
   GeneralNews = 'GeneralNews',
   EntertainmentNews = 'EntertainmentNews',
   PoliticsNews = 'PoliticsNews',
   ScienceNews = 'ScienceNews',
   MusicNews = 'MusicNews',
   SportsNews = 'SportsNews',
   TechnologyNews = 'TechnologyNews'
}